configurado projeto para TS
configurado projeto para TS e atualizado depedencias EXPO, react-native
