--- README.md
@@ -1,26 +1,0 @@
-# Ubifood 🍕🔎
-
-1. clonar repositorio:
-```bash
-git clone https://github.com/Ubifood-unipetech/Ubifood-frontend.git
-```
-
-2. abrir a pasta do projeto:
-```bash
-cd Ubifood-frontend
-```
-
-3. iniciar projeto:
-```bash
-npx expo start
-```
-
-4. digitar a letra de onde irá ser renderizado:
-- `a` for android, 
-
-- `b` [i] for ios,
-
-- `w` [w] for web.
-
-    - para ser renderizado em aparelho físico é necessario instalar o aplicativo [Expo Go](https://play.google.com/store/apps/details?id=host.exp.exponent&referrer=www) e ler o QRcode gerado pelo expo em seu terminal.
-
