+++ ios/Podfile.properties.json
@@ -1,3 +1,4 @@
 {
-  "expo.jsEngine": "jsc"
+  "expo.jsEngine": "hermes",
+  "EX_DEV_CLIENT_NETWORK_INSPECTOR": "true"
 }
