--- ios/Ubifood/AppDelegate.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <React/RCTBridgeDelegate.h>
-#import <UIKit/UIKit.h>
-
-#import <Expo/Expo.h>
-
-@interface AppDelegate : EXAppDelegateWrapper <RCTBridgeDelegate>
-
-@end
