+++ ios/barets.xcodeproj/xcshareddata/xcschemes/barets.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "Ubifood.app"
-               BlueprintName = "Ubifood"
-               ReferencedContainer = "container:Ubifood.xcodeproj">
+               BuildableName = "barets.app"
+               BlueprintName = "barets"
+               ReferencedContainer = "container:barets.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "UbifoodTests.xctest"
-               BlueprintName = "UbifoodTests"
-               ReferencedContainer = "container:Ubifood.xcodeproj">
+               BuildableName = "baretsTests.xctest"
+               BlueprintName = "baretsTests"
+               ReferencedContainer = "container:barets.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Ubifood.app"
-            BlueprintName = "Ubifood"
-            ReferencedContainer = "container:Ubifood.xcodeproj">
+            BuildableName = "barets.app"
+            BlueprintName = "barets"
+            ReferencedContainer = "container:barets.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Ubifood.app"
-            BlueprintName = "Ubifood"
-            ReferencedContainer = "container:Ubifood.xcodeproj">
+            BuildableName = "barets.app"
+            BlueprintName = "barets"
+            ReferencedContainer = "container:barets.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
