+++ ios/barets/Images.xcassets/SplashScreen.imageset/Contents.json
