+++ ios/barets/Images.xcassets/SplashScreenBackground.imageset/Contents.json
